+++ tests/testthat/test-warfarin.R
@@ -72,3 +72,4 @@ nmTest({
 
   expect_equal(run007F2$objf, run007F$objf)
 })
+})